move BugInvestigation_3F489XV0 tests: - from: MySqlMiscellaneousTests, MySqlConnectorMiscellaneousTests - to: BaseMiscellaneousTests class - model moved to Models/AccessSpecifiersPoco.cs - build script ported to all supported dbms in Scripts/ directory
