+++ PetaPoco.Tests.Integration/Scripts/MSAccessBuildDatabase.sql
@@ -98,3 +98,13 @@ CREATE TABLE [BugInvestigation_10R9LZYK] (
 	[Id] AUTOINCREMENT PRIMARY KEY,
 	[TestColumn1] BINARY
 );
+
+DROP TABLE [BugInvestigation_3F489XV0];
+
+CREATE TABLE [BugInvestigation_3F489XV0] (
+	[Id] AUTOINCREMENT PRIMARY KEY,
+	[TC1] INTEGER NOT NULL,
+	[TC2] INTEGER NOT NULL,
+	[TC3] INTEGER NOT NULL,
+	[TC4] INTEGER NOT NULL
+);
