+++ PetaPoco.Tests.Integration/Scripts/MSSQLCeBuildDatabase.sql
@@ -75,4 +75,14 @@ DROP TABLE [BugInvestigation_10R9LZYK];
 CREATE TABLE [BugInvestigation_10R9LZYK] (
 	[Id] INT IDENTITY(1,1) PRIMARY KEY,
 	[TestColumn1] BINARY(32)
-)
+);
+
+DROP TABLE [BugInvestigation_3F489XV0];
+
+CREATE TABLE [BugInvestigation_3F489XV0] (
+	[Id] INT IDENTITY(1,1) PRIMARY KEY,
+	[TC1] INT NOT NULL,
+	[TC2] INT NOT NULL,
+	[TC3] INT NOT NULL,
+	[TC4] INT NOT NULL
+);
