+++ PetaPoco.Tests.Integration/Scripts/MariaDbBuildDatabase.sql
@@ -83,6 +83,16 @@ CREATE TABLE `BugInvestigation_10R9LZYK` (
 	`TestColumn1` VARBINARY(32)
 ) ENGINE=INNODB;
 
+DROP TABLE IF EXISTS `BugInvestigation_3F489XV0`;
+
+CREATE TABLE `BugInvestigation_3F489XV0` (
+	`Id` INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	`TC1` INT NOT NULL,
+	`TC2` INT NOT NULL,
+	`TC3` INT NOT NULL,
+	`TC4` INT NOT NULL
+) ENGINE=INNODB;
+
 -- Stored Procedures
 
 DROP PROCEDURE IF EXISTS `SelectPeople`;
