+++ PetaPoco.Tests.Integration/Scripts/PostgresBuildDatabase.sql
@@ -89,6 +89,16 @@ CREATE TABLE "BugInvestigation_10R9LZYK" (
 	"TestColumn1" BYTEA
 );
 
+DROP TABLE IF EXISTS "BugInvestigation_3F489XV0";
+
+CREATE TABLE "BugInvestigation_3F489XV0" (
+	"Id" SERIAL PRIMARY KEY NOT NULL,
+	"TC1" INT NOT NULL,
+	"TC2" INT NOT NULL,
+	"TC3" INT NOT NULL,
+	"TC4" INT NOT NULL
+);
+
 DROP TABLE IF EXISTS "BugInvestigation_7K2TX4VR";
 
 CREATE TABLE "BugInvestigation_7K2TX4VR" (
