+++ PetaPoco.Tests.Integration/Scripts/SqliteBuildDatabase.sql
@@ -80,3 +80,13 @@ CREATE TABLE [BugInvestigation_10R9LZYK] (
 	[Id] INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
 	[TestColumn1] BLOB(32)
 );
+
+DROP TABLE IF EXISTS [BugInvestigation_3F489XV0];
+
+CREATE TABLE [BugInvestigation_3F489XV0] (
+	[Id] INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+	[TC1] INTEGER NOT NULL,
+	[TC2] INTEGER NOT NULL,
+	[TC3] INTEGER NOT NULL,
+	[TC4] INTEGER NOT NULL
+);
